+++ tests/api/test-all-public-symbols.c
@@ -251,6 +251,7 @@ static duk_ret_t test_func(duk_context *ctx, void *udata) {
 	(void) duk_safe_call(ctx, NULL, NULL, 0, 0);
 	(void) duk_safe_to_lstring(ctx, 0, NULL);
 	(void) duk_safe_to_string(ctx, 0);
+	(void) duk_samevalue(ctx, 0, 0);
 	(void) duk_set_finalizer(ctx, 0);
 	(void) duk_set_global_object(ctx);
 	(void) duk_set_magic(ctx, 0, 0);