+++ RELEASES.rst
@@ -1843,6 +1843,8 @@ Planned
 
 * Add support for ES6 Object.assign() (GH-1064)
 
+* Add support for ES6 Object.is() and duk_samevalue() API call (GH-1068)
+
 * Add TextEncoder and TextDecoder built-ins (the Encoding API) which allow
   Ecmascript code to read and write text stored in an ArrayBuffer or a plain
   buffer (GH-975)